Refactor friendList, added callbacks for type and call
