+++ frontend/src/Root.js
@@ -6,6 +6,7 @@ import { connect } from 'react-redux';
 import { toast } from 'react-toastify';
 
 import Content from './Content';
+import FriendsComponent from './friendList/FriendsComponent';
 
 const RootContainer = styled.div`
   display: flex;
@@ -16,7 +17,13 @@ const ConnectContainer = styled.div`
   display: flex;
   flex-direction: column;
 `;
-const FriendList = styled.div``;
+
+// ===== mock for FriendList =====
+const allUsers = ['Jarosław', 'Piotr', 'Wojciech', 'Mateusz'];
+const onFriendClick = () => alert('Typing to friend!');
+const onFriendCall = () => alert('Calling to friend!');
+// ===== end mock for FriendList =====
+
 
 const getDevices = () => new Promise((resolve) => {
   if (!navigator.mediaDevices || !navigator.mediaDevices.enumerateDevices) {
@@ -153,7 +160,11 @@ class Root extends React.Component {
           <input type="submit" value="Connect" onClick={this.onClick} />
         </ConnectContainer>
 
-        <FriendList />
+        <FriendsComponent 
+          allUsers={allUsers} 
+          onFriendClick={onFriendClick} 
+          onFriendCall={onFriendCall} 
+        />
         <Content selected={null} />
         {/* <Chat onVideo={this.onVideo} /> */}
       </RootContainer>
